pyproject.toml
[tool.poetry]
name = "django-training"
version = "0.1.0"
description = ""
authors = ["3RB16 <muhammad3rbi@gmail.com>"]
readme = "README.md"
packages = [{include = "django_training"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
djangorestframework = "^3.14.0"
django-imagekit = "^4.1.0"
pillow = "^9.2.0"
django-extensions = "^3.2.1"
django-rest-knox = "^4.2.0"
pytest-django = "^4.5.2"
django-filter = "^22.1"
celery = "^5.2.7"
django-environ = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
